chore: start using app router in the website
