+++ packages/website/.eslintrc.js
@@ -1,13 +1,14 @@
 module.exports = {
-    root: true,
-    extends: 'next',
-    'settings': {
-        'next': {
-            'rootDir': 'packages/website/'
-        }
-    },
-    'rules': {
-        '@next/next/no-img-element': 'off',
-        '@next/next/no-css-tags': 'off'
+  root: true,
+  extends: 'next',
+  'settings': {
+    'next': {
+      'rootDir': 'packages/website/'
     }
-};
+  },
+  'rules': {
+    '@next/next/no-img-element': 'off',
+    '@next/next/no-css-tags': 'off',
+    'indent': ['error', 2]
+  }
+}
