+++ packages/website/package.json
@@ -1,51 +1,52 @@
 {
-    "private": true,
-    "name": "@figma-export/website",
-    "version": "6.0.0",
-    "description": "Website for @figma-export",
-    "author": "Marco Montalbano",
-    "keywords": [
-        "figma",
-        "export",
-        "design",
-        "icons",
-        "typography",
-        "components"
-    ],
-    "homepage": "https://figma-export.marcomontalbano.com/",
-    "license": "MIT",
-    "scripts": {
-        "clean": "rimraf dist output .next out",
-        "figma:export": "figma-export use-config .figmaexportrc.mjs",
-        "next:lint": "next lint",
-        "next:dev": "next dev",
-        "next:build": "next build",
-        "dev": "run-s figma:export next:dev",
-        "build": "run-s clean figma:export next:build"
-    },
-    "devDependencies": {
-        "@figma-export/cli": "^6.0.0",
-        "@figma-export/utils": "^6.0.0",
-        "@types/react": "^18.3.1",
-        "eslint-config-next": "^14.2.3",
-        "npm-run-all": "~4.1.5",
-        "rimraf": "~5.0.6",
-        "sass": "~1.77.0"
-    },
-    "dependencies": {
-        "@svgr/plugin-jsx": "^8.1.0",
-        "next": "^14.2.3",
-        "prismjs": "^1.29.0",
-        "react": "^18.3.1",
-        "react-dom": "^18.3.1"
-    },
-    "engines": {
-        "node": ">= 18.17"
-    },
-    "browserslist": [
-        ">0.3%",
-        "not ie 11",
-        "not dead",
-        "not op_mini all"
-    ]
+  "private": true,
+  "name": "@figma-export/website",
+  "version": "6.0.0",
+  "description": "Website for @figma-export",
+  "author": "Marco Montalbano",
+  "keywords": [
+    "figma",
+    "export",
+    "design",
+    "icons",
+    "typography",
+    "components"
+  ],
+  "homepage": "https://figma-export.marcomontalbano.com/",
+  "license": "MIT",
+  "scripts": {
+    "clean": "rimraf dist output .next out",
+    "figma:export": "figma-export use-config .figmaexportrc.mjs",
+    "next:lint": "next lint",
+    "next:dev": "next dev",
+    "next:build": "next build",
+    "dev": "run-s figma:export next:dev",
+    "build": "run-s clean figma:export next:build"
+  },
+  "devDependencies": {
+    "@figma-export/cli": "^6.0.0",
+    "@figma-export/utils": "^6.0.0",
+    "@types/react": "^18.3.1",
+    "eslint-config-next": "^14.2.3",
+    "npm-run-all": "~4.1.5",
+    "rimraf": "~5.0.6",
+    "sass": "~1.77.0"
+  },
+  "dependencies": {
+    "@svgr/plugin-jsx": "^8.1.0",
+    "@types/prismjs": "^1.26.4",
+    "next": "^14.2.3",
+    "prismjs": "^1.29.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
+  },
+  "engines": {
+    "node": ">= 18.17"
+  },
+  "browserslist": [
+    ">0.3%",
+    "not ie 11",
+    "not dead",
+    "not op_mini all"
+  ]
 }
