+++ packages/website/src/app/layout.tsx
@@ -1,0 +1,40 @@
+import 'prismjs/themes/prism-solarizedlight.min.css'
+import '../../scss/index.scss'
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+    children: React.ReactNode
+}>) {
+  return (
+    <html lang="en">
+      <head>
+        <meta charSet="utf-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>@figma-export</title>
+        <meta name="description" content="Export tool for Figma. You can easily and automatically export your figma components and use them directly into your website." />
+        <link rel="apple-touch-icon" sizes="57x57" href="/images/icon/apple-icon-57x57.png" />
+        <link rel="apple-touch-icon" sizes="60x60" href="/images/icon/apple-icon-60x60.png" />
+        <link rel="apple-touch-icon" sizes="72x72" href="/images/icon/apple-icon-72x72.png" />
+        <link rel="apple-touch-icon" sizes="76x76" href="/images/icon/apple-icon-76x76.png" />
+        <link rel="apple-touch-icon" sizes="114x114" href="/images/icon/apple-icon-114x114.png" />
+        <link rel="apple-touch-icon" sizes="120x120" href="/images/icon/apple-icon-120x120.png" />
+        <link rel="apple-touch-icon" sizes="144x144" href="/images/icon/apple-icon-144x144.png" />
+        <link rel="apple-touch-icon" sizes="152x152" href="/images/icon/apple-icon-152x152.png" />
+        <link rel="apple-touch-icon" sizes="180x180" href="/images/icon/apple-icon-180x180.png" />
+        <link rel="icon" type="image/png" sizes="192x192"  href="/images/icon/android-icon-192x192.png" />
+        <link rel="icon" type="image/png" sizes="32x32" href="/images/icon/favicon-32x32.png" />
+        <link rel="icon" type="image/png" sizes="96x96" href="/images/icon/favicon-96x96.png" />
+        <link rel="icon" type="image/png" sizes="16x16" href="/images/icon/favicon-16x16.png" />
+        <link rel="manifest" href="/images/icon/manifest.json" />
+        <link href="//fonts.googleapis.com/css2?family=Lato:ital,wght@0,300;0,700;1,300;1,700&family=Roboto&family=Roboto+Condensed:wght@400;700&family=Spinnaker&display=swap" rel="stylesheet" />
+        <meta name="msapplication-TileColor" content="#ffffff" />
+        <meta name="msapplication-TileImage" content="/images/icon/ms-icon-144x144.png" />
+        <meta name="theme-color" content="#ffffff" />
+      </head>
+      <body>
+        {children}
+      </body>
+    </html>
+  )
+}
