+++ packages/website/src/components/output-styles/index.tsx
@@ -1,0 +1,10 @@
+import FigmaStyles from './FigmaStyles';
+
+const OutputStyles = () => (
+  <div>
+    <h2 id="figma-styles" className="figma-gradient with-opacity-10">Figma Styles</h2>
+    <FigmaStyles />
+  </div>
+);
+
+export default OutputStyles;
